// src/webkit-greeter.d.ts
type Session = {
  key: string;
  name: string;
  type: string;
  comment: string;
};

type User = {
  display_name: string;
  home_directory: string;
  image: string;
  language: string;
  session: string;
  username: string;
  // lightdm only
  background?: string | null;
  logged_in?: boolean;
  layout?: string | null;
  layouts?: string[] | null;
};